+++ app/api/api/urls.py
@@ -2,6 +2,7 @@ from api import views
 from api.deprecated_router import router as deprecated_router
 from django.urls import include, path
 from shared.files.views import FileDownloadView
+from shared.jobs.views import JobView
 
 from .deprecated_urls import urlpatterns as deprecated_urlpatterns
 
@@ -29,7 +30,7 @@ urlpatterns = [
     ),
     path(
         "v2/scanreports/<int:pk>/jobs/",
-        views.ScanReportJobView.as_view(),
+        JobView.as_view(),
         name="scan-report-jobs",
     ),
     path(
@@ -80,6 +81,11 @@ urlpatterns = [
         "v2/scanreports/<int:pk>/tables/<int:table_pk>/",
         views.ScanReportTableDetailV2.as_view(),
         name="scan-report-table-detail",
+    ),
+    path(
+        "v2/scanreports/<int:pk>/tables/<int:table_pk>/jobs/",
+        JobView.as_view(),
+        name="scan-report-table-jobs",
     ),
     path(
         "v2/scanreports/<int:pk>/tables/<int:table_pk>/fields/",
